refactor based on review feedback
Moves sort_objects to a new sort_objects.py.  Misc other minor refactors.
