+++ lightkube/__init__.py
@@ -1,4 +1,5 @@
 from .core.client import Client, AsyncClient
 from .core.generic_client import ALL_NS
 from .core.exceptions import ApiError, ConfigError, LoadResourceError
+from .core.sort_objects import sort_objects
 from .config.kubeconfig import KubeConfig, SingleConfig
